requirements.txt
qdrant-client==1.16.2
sentence-transformers>=2.2.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
numpy>=2.3.0